+++ src/gui/text/textrenderer.cpp
@@ -62,7 +62,7 @@ void TextRenderer::renderText( QString text, int x, int y, int width, int height
     program->bind();
 
     // Offset for position
-    long int offset = 0;
+    intptr_t offset = 0;
     // Tell OpenGL programmable pipeline how to locate vertex position data
     int vertexLocation = program->attributeLocation( "a_position" );
     program->enableAttributeArray( vertexLocation );
